循環importを修正
