+++ src/func.ts
@@ -1,65 +1,15 @@
 import { Member } from "./member.js";
 import { valType } from "./message.js";
-import { Field, FieldBase } from "./field.js";
+import { Field } from "./field.js";
 import * as Message from "./message.js";
+import {
+  Arg,
+  FuncCallback,
+  FuncInfo,
+  FuncPromiseData,
+  Val,
+} from "./funcBase.js";
 
-export type Val = string | number | boolean;
-
-export interface FuncInfo {
-  returnType: number;
-  args: Arg[];
-  funcImpl?: FuncCallback;
-  // call
-}
-
-export interface Arg {
-  name?: string;
-  type?: number;
-  init?: Val | null;
-  min?: number | null;
-  max?: number | null;
-  option?: string[] | number[];
-}
-
-export class FuncNotFoundError extends Error {
-  constructor(base: FieldBase) {
-    super(`member("${base.member_}").func("${base.field_}") is not set`);
-    this.name = "FuncNotFoundError";
-  }
-}
-
-export class FuncPromiseData {
-  callerId: number;
-  caller: string;
-  reach: Promise<boolean>;
-  resolveReach: (r: boolean) => void = () => undefined;
-  finish: Promise<Val>;
-  resolveFinish: (r: Val | Promise<Val>) => void = () => undefined;
-  // 例外をセットする
-  rejectFinish: (e: Error) => void = () => undefined;
-  base: Field;
-
-  constructor(callerId: number, caller: string, base: Field) {
-    this.base = base;
-    this.callerId = callerId;
-    this.caller = caller;
-    this.reach = new Promise((res) => {
-      this.resolveReach = (r: boolean) => {
-        res(r);
-        if (!r) {
-          this.rejectFinish(new FuncNotFoundError(this.base));
-        }
-      };
-    });
-    this.finish = new Promise((res, rej) => {
-      this.resolveFinish = res;
-      this.rejectFinish = rej;
-    });
-  }
-  getter() {
-    return new FuncPromise(this);
-  }
-}
 /**
  * 非同期で実行した関数の実行結果を表す。
  */
@@ -94,11 +44,7 @@ export class FuncPromise {
    */
   result: Promise<Val>;
   constructor(pData: FuncPromiseData) {
-    this.base_ = new Field(
-      pData.base.data,
-      pData.base.member_,
-      pData.base.field_
-    );
+    this.base_ = new Field(pData.data, pData.base.member_, pData.base.field_);
     this.reach = this.started = pData.reach;
     this.finish = this.result = pData.finish;
   }
@@ -148,8 +94,6 @@ export function runFunc(fi: FuncInfo, args: Val[]) {
     );
   }
 }
-
-export type FuncCallback = (...args: any[]) => Val | Promise<Val> | void;
 
 /**
  * Funcを指すクラス
@@ -272,11 +216,13 @@ export class Func {
    * 戻り値やエラー、例外はFuncPromiseから取得する
    */
   runAsync(...args: Val[]) {
-    const r = this.base_.dataCheck().funcResultStore.addResult("", this.base_);
+    const r = this.base_
+      .dataCheck()
+      .funcResultStore.addResult("", this.base_, this.base_.dataCheck());
     setTimeout(() => {
       this.runImpl(r, args);
     });
-    return r.getter();
+    return new FuncPromise(r);
   }
   /**
    * 関数が存在すればtrueを返す
