+++ src/logBase.ts
@@ -1,0 +1,5 @@
+export interface LogLine {
+  level: number;
+  time: Date;
+  message: string;
+}
