+++ test/client.spec.ts
@@ -8,14 +8,15 @@ import { Text } from "../src/text.js";
 import { RobotModel, RobotLink } from "../src/robotModel.js";
 import { Transform } from "../src/transform.js";
 import { Geometry } from "../src/canvas3d.js";
-import { Func, FuncNotFoundError, FuncPromiseData } from "../src/func.js";
+import { Func, FuncPromise } from "../src/func.js";
+import { FuncNotFoundError, FuncPromiseData } from "../src/funcBase.js";
 import { valType } from "../src/message.js";
 import { View, viewComponents } from "../src/view.js";
-import { Field } from "../src/field.js";
 import { Member } from "../src/member.js";
 import { eventType } from "../src/event.js";
 import version from "../src/version.js";
 import { Log } from "../src/log.js";
+import { FieldBase } from "../src/fieldBase.js";
 
 describe("Client Tests", function () {
   const selfName = "test";
@@ -1069,7 +1070,7 @@ describe("Client Tests", function () {
         wcli.start();
         setTimeout(() => {
           data.memberIds.set("a", 10);
-          const r = new FuncPromiseData(1, "", new Field(data, "a", "b"));
+          const r = new FuncPromiseData(1, "", new FieldBase("a", "b"), data);
           data.pushSendAlways([
             {
               kind: Message.kind.call,
@@ -1101,7 +1102,8 @@ describe("Client Tests", function () {
           data.memberIds.set("a", 10);
           const r = data.funcResultStore.addResult(
             "",
-            new Field(data, "a", "b")
+            new FieldBase("a", "b"),
+            data
           );
           assert.strictEqual(r.callerId, 0);
           data.pushSendAlways([
@@ -1121,11 +1123,11 @@ describe("Client Tests", function () {
               c: 0,
               s: false,
             });
-            r.getter()
-              .started.then((started) => assert.isFalse(started))
+            new FuncPromise(r).started
+              .then((started) => assert.isFalse(started))
               .catch(() => assert.fail("r.started threw error"));
-            r.getter()
-              .result.then(() => {
+            new FuncPromise(r).result
+              .then(() => {
                 assert.fail("r.result did not throw error");
                 done();
               })
@@ -1142,7 +1144,8 @@ describe("Client Tests", function () {
           data.memberIds.set("a", 10);
           const r = data.funcResultStore.addResult(
             "",
-            new Field(data, "a", "b")
+            new FieldBase("a", "b"),
+            data
           );
           data.pushSendAlways([
             {
@@ -1168,11 +1171,11 @@ describe("Client Tests", function () {
               e: true,
               r: "aaa",
             });
-            r.getter()
-              .started.then((started) => assert.isFalse(started))
+            new FuncPromise(r).started
+              .then((started) => assert.isFalse(started))
               .catch(() => assert.fail("r.started threw error"));
-            r.getter()
-              .result.then(() => {
+            new FuncPromise(r).result
+              .then(() => {
                 assert.fail("r.result did not throw error");
                 done();
               })
@@ -1189,7 +1192,8 @@ describe("Client Tests", function () {
           data.memberIds.set("a", 10);
           const r = data.funcResultStore.addResult(
             "",
-            new Field(data, "a", "b")
+            new FieldBase("a", "b"),
+            data
           );
           data.pushSendAlways([
             {
@@ -1215,11 +1219,11 @@ describe("Client Tests", function () {
               e: false,
               r: "aaa",
             });
-            r.getter()
-              .started.then((started) => assert.isTrue(started))
+            new FuncPromise(r).started
+              .then((started) => assert.isTrue(started))
               .catch(() => assert.fail("r.started threw error"));
-            r.getter()
-              .result.then((res) => {
+            new FuncPromise(r).result
+              .then((res) => {
                 assert.strictEqual(res, "aaa");
                 done();
               })
